requirements.txt
alembic==0.8.9
aniso8601==1.2.0
blinker==1.4
click==6.7
colorama==0.3.7
coverage==4.3.1
Flask==0.12
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
Flask-WhooshAlchemy==0.56
itsdangerous==0.24
Jinja2==2.9.3
Mako==1.0.6
MarkupSafe==0.23
marshmallow==2.11.1
nose==1.3.7
PyJWT==1.4.2
python-dateutil==2.6.0
python-editor==1.0.3
python-termstyle==0.1.10
pytz==2016.10
rednose==1.2.1
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.15
Whoosh==2.7.4
